fix unstable app settings reference
