+++ frontend/components/games/games-page.tsx
@@ -1,27 +1,14 @@
 import { Group, Stack } from "@mantine/core";
-import { useCallback, useEffect, useRef } from "react";
 import { FilterMenu } from "@components/filters/filter-menu";
 import { RefreshButton } from "@components/refresh-button";
 import { AddGame } from "./add-game-button";
 import { useAppEvent } from "@hooks/use-app-event";
-import { GamesSortBy } from "@api/bindings";
-import { useAtomValue, useSetAtom } from "jotai";
-import { gameDataAtom } from "@hooks/use-data";
+import { useSetAtom } from "jotai";
 import { selectedGameAtom } from "./games-state";
-import { TableContainer } from "@components/table/table-container";
-import { TableVirtuoso, TableVirtuosoHandle } from "react-virtuoso";
-import { useVirtuosoHeaderContent } from "@hooks/use-virtuoso-header-content";
-import { useVirtuosoTableComponents } from "@hooks/use-virtuoso-table-components";
-import { GameRow, gameRowHeight } from "./game-row";
-import { useDataQuery } from "@hooks/use-data-query";
-import { gamesColumns } from "./games-columns";
-import styles from "./games.module.css";
+import { GamesTable } from "./games-table";
 
 export function GamesPage() {
-	const gameData = useAtomValue(gameDataAtom);
 	const setSelectedGame = useSetAtom(selectedGameAtom);
-	const [dataQuery, setDataQuery] = useDataQuery();
-	const tableRef = useRef<TableVirtuosoHandle>(null);
 
 	useAppEvent("selectInstalledGame", "games-page", ([providerId, gameId]) => {
 		setSelectedGame({
@@ -30,34 +17,6 @@ export function GamesPage() {
 		});
 	});
 
-	const onChangeSort = useCallback(
-		(sortBy: GamesSortBy) => {
-			const sortDescending =
-				sortBy === dataQuery?.sortBy && !dataQuery?.sortDescending;
-
-			setDataQuery({
-				sortBy,
-				sortDescending,
-			});
-		},
-		[dataQuery, setDataQuery],
-	);
-
-	const renderHeaders = useVirtuosoHeaderContent(
-		gamesColumns,
-		onChangeSort,
-		dataQuery?.sortBy,
-		dataQuery?.sortDescending,
-	);
-
-	const tableComponents = useVirtuosoTableComponents(GameRow);
-
-	useEffect(() => {
-		if (tableRef.current) {
-			tableRef.current.scrollToIndex(0);
-		}
-	}, [dataQuery]);
-
 	return (
 		<Stack h="100%">
 			<Group>
@@ -65,21 +24,7 @@ export function GamesPage() {
 				<FilterMenu />
 				<RefreshButton />
 			</Group>
-			<TableContainer>
-				<TableVirtuoso
-					ref={tableRef}
-					className={styles.table}
-					components={tableComponents}
-					fixedHeaderContent={renderHeaders}
-					data={gameData.gameIds}
-					fixedItemHeight={gameRowHeight}
-					overscan={50}
-					increaseViewportBy={100}
-					computeItemKey={(index) =>
-						`${gameData.gameIds[index]?.providerId}${gameData.gameIds[index]?.gameId}`
-					}
-				/>
-			</TableContainer>
+			<GamesTable />
 		</Stack>
 	);
 }
