+++ frontend/hooks/use-app-settings.ts
@@ -3,33 +3,38 @@ import { atom, useAtom } from "jotai";
 import { useCallback, useEffect } from "react";
 
 const defaultSettings: AppSettings = {
-  hideGameThumbnails: false,
-  overrideLanguage: null,
+	hideGameThumbnails: false,
+	overrideLanguage: null,
 };
 
-const appSettingsAtom = atom(defaultSettings);
+const appSettingsAtom = atom({
+	isInitialized: false,
+	settings: defaultSettings,
+});
 
 export function useAppSettings() {
-  const [settings, setSettingsInternal] = useAtom(appSettingsAtom);
+	const [settings, setSettingsInternal] = useAtom(appSettingsAtom);
 
-  useEffect(() => {
-    commands.getAppSettings().then((result) => {
-      if (result.status === "ok") {
-        setSettingsInternal(result.data);
-      }
-    });
-  }, [setSettingsInternal]);
+	useEffect(() => {
+		if (settings.isInitialized) return;
 
-  const setSettings = useCallback(
-    (newSettings: AppSettings) => {
-      commands.saveAppSettings(newSettings).then((result) => {
-        if (result.status === "ok") {
-          setSettingsInternal(newSettings);
-        }
-      });
-    },
-    [setSettingsInternal],
-  );
+		commands.getAppSettings().then((result) => {
+			if (result.status === "ok") {
+				setSettingsInternal({ isInitialized: true, settings: result.data });
+			}
+		});
+	}, [setSettingsInternal, settings]);
 
-  return [settings, setSettings] as const;
+	const setSettings = useCallback(
+		(newSettings: AppSettings) => {
+			commands.saveAppSettings(newSettings).then((result) => {
+				if (result.status === "ok") {
+					setSettingsInternal({ isInitialized: true, settings: newSettings });
+				}
+			});
+		},
+		[setSettingsInternal],
+	);
+
+	return [settings.settings, setSettings] as const;
 }
